+++ NEWS.rst
@@ -1,3 +1,10 @@
+Changes in 3.7.2 (released 02/26/2010)
+--------------------------------------
+
+*	Fixed a bug in the handling of the ``addlist`` opcode in
+	:func:`ll.xist.ns.jsp.fromul4`.
+
+
 Changes in 3.7.1 (released 02/08/2010)
 --------------------------------------